+++ kernel/sched/sched_i.h
@@ -18,7 +18,7 @@ void runqueue_init(void);
 void ctx_switch_full(volatile void* next_ctx, volatile void* cur_ctx);
 void ctx_switch_half(volatile void* next_ctx) __attribute__((noreturn));
 
-tcb_t* _get_runList_tcb(int prio);
+tcb_t* _get_runList_tcb(uint8_t prio);
 
 /** A list of every TCB+kernel stack that can exist on this system.
  * Note -- this list is not ordered by anything.
